+++ linalg/bwc/mpfq/mpfq_pz.hpp
@@ -152,8 +152,8 @@ std::ostream& mpfq_pz_cxx_out(mpfq_pz_dst_field, std::ostream&, mpfq_pz_src_elt)
 std::istream& mpfq_pz_cxx_in(mpfq_pz_dst_field, std::istream&, mpfq_pz_dst_elt);
 
 /* Vector functions */
-std::ostream& mpfq_pz_vec_cxx_out(mpfq_pz_dst_field, std::ostream&, mpfq_pz_src_vec, unsigned int);
-std::istream& mpfq_pz_vec_cxx_in(mpfq_pz_dst_field, std::istream&, mpfq_pz_vec *, unsigned int *);
+std::ostream& mpfq_pz_vec_cxx_out(mpfq_pz_dst_field, std::ostream&, mpfq_pz_src_vec, unsigned long);
+std::istream& mpfq_pz_vec_cxx_in(mpfq_pz_dst_field, std::istream&, mpfq_pz_vec *, unsigned long *);
 
 /* Polynomial functions */
 /* *Mpfq::defaults::poly::code_for_poly_cxx_out, pz */
